+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Payload formatter documentation.
 - CLI support for setting message payload formatters from a local file. (see `--formatters.down-formatter-parameter-local-file` and `--formatters.up-formatter-parameter-local-file` options).
 - Generic encryption/decryption to KeyVault.
+- Option to ignore log messages for selected gRPC method on success (see `grpc.log-ignore-methods` option).
 
 ### Changed